+++ zerver/tests/test_auth_backends.py
@@ -40,13 +40,13 @@ from social_django.storage import BaseDjangoStorage
 from social_django.strategy import DjangoStrategy
 
 from confirmation.models import Confirmation, create_confirmation_link
+from zerver.actions.invites import do_invite_users
 from zerver.lib.actions import (
     change_user_is_active,
     do_create_realm,
     do_create_user,
     do_deactivate_realm,
     do_deactivate_user,
-    do_invite_users,
     do_reactivate_realm,
     do_reactivate_user,
     do_set_realm_property,
